requirements.txt
async-timeout      4.0.2
Authlib            1.0.0
beautifulsoup4     4.10.0
bidict             0.22.0
bs4                0.0.1
certifi            2021.10.8
cffi               1.15.0
chardet            3.0.4
charset-normalizer 2.0.12
click              8.0.4
colorama           0.4.4
cryptography       36.0.2
Deprecated         1.2.13
distlib            0.3.4
filelock           3.6.0
Flask              2.0.2
Flask-Cors         3.0.10
Flask-Environ      0
Flask-Ext          0.1
Flask-Login        0.4.1
Flask-OAuth        0.12
Flask-SocketIO     5.2.0
gunicorn           20.1.0
httplib2           0.20.4
idna               2.8
importlib-metadata 4.11.3
itsdangerous       2.1.1
Jinja2             3.0.3
lxml               4.8.0
MarkupSafe         2.1.1
mod-wsgi           4.9.1
oauth2             1.9.0.post1
oauthlib           3.0.1
packaging          21.3
pip                21.2.4
platformdirs       2.5.1
psycopg2           2.9.3
pycparser          2.21
PyJWT              1.4.2
pyOpenSSL          22.0.0
pyparsing          3.0.8
python-dotenv      0.19.2
python-engineio    4.3.2
python-http-client 3.3.7
python-socketio    5.6.0
redis              4.3.1
requests           2.21.0
setuptools         3.3
six                1.16.0
soupsieve          2.3.1
starkbank-ecdsa    2.0.3
urllib3            1.24.3
virtualenv         20.13.2
Werkzeug           2.0.0
wrapt              1.14.1
zipp               3.8.0